refactor(java): Move the AuthorizationManager to the common package.
