+++ src/main/java/com/synopsys/integration/alert/web/audit/AuditEntryController.java
@@ -42,9 +42,9 @@ import com.synopsys.integration.alert.common.exception.AlertNotificationPurgedEx
 import com.synopsys.integration.alert.common.persistence.model.AuditEntryModel;
 import com.synopsys.integration.alert.common.persistence.model.AuditJobStatusModel;
 import com.synopsys.integration.alert.common.rest.model.AlertPagedModel;
+import com.synopsys.integration.alert.common.security.authorization.AuthorizationManager;
 import com.synopsys.integration.alert.web.controller.BaseController;
 import com.synopsys.integration.alert.web.controller.ResponseFactory;
-import com.synopsys.integration.alert.web.security.authorization.AuthorizationManager;
 import com.synopsys.integration.exception.IntegrationException;
 
 @RestController
